'feature-search-speed-history-progressbarupload-converttomp3-changeicon'

+++ app/src/main/java/com/example/duan1_customer/model/Request_Update_Minute_Tele.java
@@ -2,12 +2,12 @@ package com.example.duan1_customer.model;
 
 public class Request_Update_Minute_Tele {
     public String telesale_id;
-    public float minutes;
+    public double minutes;
     public String service;
     public String date;
     public boolean missed;
 
-    public Request_Update_Minute_Tele(String telesale_id, float minutes, String service, String date, boolean missed) {
+    public Request_Update_Minute_Tele(String telesale_id, double minutes, String service, String date, boolean missed) {
         this.telesale_id = telesale_id;
         this.minutes = minutes;
         this.service = service;
@@ -15,7 +15,7 @@ public class Request_Update_Minute_Tele {
         this.missed = missed;
     }
 
-    public Request_Update_Minute_Tele(String telesale_id, float minutes, String service, String date) {
+    public Request_Update_Minute_Tele(String telesale_id, double minutes, String service, String date) {
         this.telesale_id = telesale_id;
         this.minutes = minutes;
         this.service = service;
